Inline RegisterHardcodedCommand call
